Major bugfix for codes handling summary-only data. The original binning codes were incorrect and new codes introduced to display idle time on the Overview tool failed to take into account the fact that idle time does not appear in summary-only data.

+++ projections/gui/Analysis.java
@@ -534,9 +534,8 @@ public class Analysis {
      */
     public static void loadSummaryData() {
 	if (sts.hasSumFiles()) { 
-	    long sizeInt = sumAnalyzer.getIntervalSize();
+	    int sizeInt=(int)(sumAnalyzer.getIntervalSize());
 	    int nInt=(int)(getTotalTime()/sizeInt);
-	    
 	    loadSummaryData(sizeInt, 0, nInt-1);
 	} else if (sts.hasSumAccumulatedFile()) {
 	    // do nothing **HACK** - action taken later.
